+++ src/javascript/7000-rendering.js
@@ -1,4 +1,3 @@
-st.page = new s.Page();
 s.menuBar = new s.ActionList();
 
 s.menuMarkup = function() {